okay a lot to type here Important things: removed assert from all the functions so it won't destroy your script if it gets a wrong value added a cWrap around the loading character so that it won't loop forever if your game doesn't load the character
Also added a return table for all the local values inside the universal script so you could idk mess around with it or somehting

main reason i did that was so that you could do a repeat until Uni.charLoading == false so that you could still wait for your character to load if you want it to.

rewrote genName again so that it uses string.char also it won't error if you don't

made some more functions also more efficent like onLeave and onCharacterLoaded so that it loops through the tables and fires them one by one instead of making a connection for each one.

removed the one second wait for when its waiting for the game to load so it will load faster

This will now return a table with all the connections it contains, the savetable that will save when you leave the game, the tables for the onload and onleave functions and the charloading boolean

anyways pretty good update for this script lol