fixed a few things and reordered things
ill update it eventually to load all the ui libs and other libs like customcmds and no lag and ultranolag

so it will loadstring all of them and put them into a function called like libraryLoad() and you just pass the library name you want to load and it will execute it. Simple, quick, neater, and more professional. Very cool

<3