added a faster and better and actually working method of loading the save file for the universal script, that means i did add a new function which is going to be local
the reason im keeping it local is because i don't want people to be using this lib to use this function, its very insecure and shouldn't be a thing that you're doing.

also copyOver function now returns the "to" table. So if you want to quickly copy a table it would look like

return copyOver(table)
or
return copyOver(table, {})

either one will work. I need to eventually create a documentation or something because there are actually a lot of functions and people might be interesting in using this library to help people code or just use useful functions

i started this lib cuz i wanted a quick and easy way to load all the functions and stuff like humanoidrootpart and players really quick without having to type it all out in the new script

it has gotten quick big and now im literally typing out actual messages in my commits so that if people are interested they can see what im doing n shit like that